+++ src/tests/integration/global_teardown.ts
@@ -1,4 +1,4 @@
-import { SharedTestServer } from './shared-server';
+import { SharedTestServer } from './shared_server';
 
 /**
  * Global teardown for integration tests
